backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
anthropic>=0.25.0
openpyxl>=3.1.2
olefile>=0.47
pydantic>=2.0.0
python-dotenv>=1.0.0
